Connection and import refactoring
