+++ server/server-sql-util.js
@@ -483,70 +483,16 @@ function setCategories (dataset, facet) {
     io.syncFacets(dataset);
   });
 }
-
 /**
  * Scan dataset and create Facets
  * when done, send new facets to client.
- *
- * Identification of column (facet) type is done by querying the postgres metadata
- * dataTypeID: 1700,         numeric
- * dataTypeID: 20, 21, 23,   integers
- * dataTypeID: 700, 701,     float8
  *
  * @function
  */
 function scanData (dataset) {
   var query = squel.select().distinct().from(esc(dataset.databaseTable)).limit(50);
-
   utilPg.queryAndCallBack(query, function (data) {
-    // remove previous facets
-    dataset.facets.reset();
-
-    data.fields.forEach(function (field) {
-      var type;
-      var subtype;
-
-      var SQLtype = field.dataTypeID;
-      if (SQLtype === 1700 || SQLtype === 20 || SQLtype === 21 || SQLtype === 23 || SQLtype === 700 || SQLtype === 701) {
-        type = 'continuous';
-      } else if (SQLtype === 17) {
-        // ignore:
-        // 17: wkb_geometry
-        console.warn('Ignoring column of type 17 (wkb_geometry)');
-        return;
-      } else if (utilPg.SQLDatetimeTypes.indexOf(SQLtype) > -1) {
-        // console.log('found: ', SQLtype);
-        type = 'timeorduration';
-        if (SQLtype === 1186) {
-          subtype = 'duration';
-        } else {
-          subtype = 'datetime';
-        }
-      } else {
-        // default to categorial
-        // console.warn('Defaulting to categorial type for SQL column type ', SQLtype);
-        type = 'categorial';
-      }
-
-      var sample = [];
-      data.rows.forEach(function (row) {
-        if (sample.length < 6 && sample.indexOf(row[field.name]) === -1) {
-          sample.push(row[field.name]);
-        }
-      });
-
-      var facet = dataset.facets.add({
-        name: field.name,
-        accessor: field.name,
-        type: type,
-        description: sample.join(', ')
-      });
-      if (facet.isTimeOrDuration) {
-        facet.timeTransform.type = subtype;
-      }
-    });
-
-    // send facets to client
+    utilPg.parseRows(data, dataset);
     io.syncFacets(dataset);
   });
 }
